+++ ScriptResolverService/CustomFunctions/F.cs
@@ -14,10 +14,22 @@ namespace EddiScriptResolverService.CustomFunctions
         public FunctionCategory Category => FunctionCategory.Utility;
         public string description => Properties.CustomFunctions_Untranslated.F;
         public Type ReturnType => typeof( string );
-        public IFunction function => Function.CreateNative1( ( runtime, scriptName, writer ) =>
+        public IFunction function => Function.CreateNativeMinMax( ( runtime, values, writer ) =>
         {
-            var result = scriptName.AsString;
-            return ScriptResolver.resolveFromName( result, Scripts, GetContext( runtime.Globals ), false )?.Trim();
-        });
+            var scriptName = values[ 0 ].AsString;
+            if ( values.Count > 1 && values[ 1 ].Type != ValueContent.Map )
+            {
+                throw new ArgumentException ( $"The function invoking {scriptName} has arguments which are not a map value." );
+            }
+
+            return ScriptInvoker.ResolveFromName(
+                scriptName,
+                Scripts,
+                GetContext( runtime.Globals ),
+                false,
+                false,
+                values.Count > 1,
+                values.Count > 1 ? values[ 1 ] : default )?.Trim();
+        }, 1, 2 );
     }
 }